Never fix bugs on the phone!
git-svn-id: https://subversion.assembla.com/svn/sdrsharp/trunk@496 d3cde5d6-82b8-4669-8b62-43054d6781d8

+++ SDRSharp/MainForm.cs
@@ -634,8 +634,8 @@ namespace SDRSharp
             spectrumAnalyzer.SpectrumWidth = (int) _streamControl.SampleRate;
             waterfall.SpectrumWidth = spectrumAnalyzer.SpectrumWidth;
 
-            frequencyNumericUpDown.Maximum = (long) (centerFreqNumericUpDown.Value + _streamControl.SampleRate / 2);
-            frequencyNumericUpDown.Minimum = (long) (centerFreqNumericUpDown.Value - _streamControl.SampleRate / 2);
+            frequencyNumericUpDown.Maximum = (long) centerFreqNumericUpDown.Value + (int) (_streamControl.SampleRate / 2);
+            frequencyNumericUpDown.Minimum = (long) centerFreqNumericUpDown.Value - (int) (_streamControl.SampleRate / 2);
 
             if (centerFreqNumericUpDown.Value != oldCenterFrequency)
             {
